notebooks/embeddings/pyproject.toml
[tool.poetry]
name = "embeddings"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.25.0"
fsspec = "^2023.6.0"
s3fs = "^2023.6.0"
python-dotenv = "^1.0.0"
openai = { version = "^0.27.8", extras = ["embeddings"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
